# docs/api.openapi.yaml
openapi: 3.0.3
info:
  title: BoxPilot API
  version: 0.1.0
  description: Control plane for sing-box. RPC-style GET/POST only.
servers:
  - url: /

tags:
  - name: System
  - name: Subscriptions
  - name: Nodes
  - name: Runtime
  - name: Settings

paths:

  /healthz:
    get:
      tags: [System]
      summary: Liveness probe
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
                example: ok

  /api/v1/runtime/status:
    get:
      tags: [Runtime]
      summary: Get runtime status
      responses:
        '200':
          description: Runtime status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeStatusResponse'

  /api/v1/runtime/traffic:
    get:
      tags: [Runtime]
      summary: Get runtime traffic snapshot
      responses:
        '200':
          description: Traffic snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeTrafficResponse'

  /api/v1/runtime/connections:
    get:
      tags: [Runtime]
      summary: Get runtime forwarding connections snapshot
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Connections snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeConnectionsResponse'

  /api/v1/runtime/logs:
    get:
      tags: [Runtime]
      summary: Get runtime diagnostic logs
      parameters:
        - name: level
          in: query
          required: false
          schema:
            type: string
            enum: [all, info, warn, error]
            default: all
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 80
      responses:
        '200':
          description: Runtime logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeLogsResponse'

  /api/v1/runtime/proxy/check:
    post:
      tags: [Runtime]
      summary: Check HTTP/SOCKS proxy chain connectivity
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuntimeProxyCheckRequest'
      responses:
        '200':
          description: Proxy check result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeProxyCheckResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'

  /api/v1/runtime/plan:
    post:
      tags: [Runtime]
      summary: Dry-run build config
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuntimePlanRequest'
      responses:
        '200':
          description: Plan result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimePlanResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'

  /api/v1/runtime/reload:
    post:
      tags: [Runtime]
      summary: Build config and restart sing-box
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuntimeReloadRequest'
      responses:
        '200':
          description: Reload result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeReloadResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'

  /api/v1/settings/forwarding/status:
    get:
      tags: [Settings]
      summary: Get forwarding runtime status
      responses:
        '200':
          description: Forwarding runtime status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForwardingRuntimeStatusResponse'

  /api/v1/settings/forwarding/summary:
    get:
      tags: [Settings]
      summary: Get forwarding runtime summary and selected nodes
      responses:
        '200':
          description: Forwarding summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForwardingSummaryResponse'

  /api/v1/settings/forwarding/policy:
    get:
      tags: [Settings]
      summary: Get forwarding node eligibility policy
      responses:
        '200':
          description: Forwarding policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForwardingPolicyResponse'

  /api/v1/settings/forwarding/policy/update:
    post:
      tags: [Settings]
      summary: Update forwarding node eligibility policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateForwardingPolicyRequest'
      responses:
        '200':
          description: Forwarding policy updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForwardingPolicyResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'

  /api/v1/settings/forwarding/start:
    post:
      tags: [Settings]
      summary: Start forwarding runtime
      responses:
        '200':
          description: Forwarding started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForwardingRuntimeStatusResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'

  /api/v1/settings/forwarding/stop:
    post:
      tags: [Settings]
      summary: Stop forwarding runtime
      responses:
        '200':
          description: Forwarding stopped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForwardingRuntimeStatusResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'

  /api/v1/settings/routing:
    get:
      tags: [Settings]
      summary: Get routing bypass settings
      responses:
        '200':
          description: Routing bypass settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutingSettingsResponse'

  /api/v1/settings/routing/update:
    post:
      tags: [Settings]
      summary: Update routing bypass settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoutingSettingsRequest'
      responses:
        '200':
          description: Routing bypass settings updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutingSettingsResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'

  /api/v1/settings/routing/summary:
    get:
      tags: [Settings]
      summary: Get routing and geo summary
      responses:
        '200':
          description: Routing diagnostics summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutingSummaryResponse'

  /api/v1/subscriptions:
    get:
      tags: [Subscriptions]
      summary: List subscriptions
      responses:
        '200':
          description: Subscription list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Subscription'
                required: [data]

  /api/v1/subscriptions/create:
    post:
      tags: [Subscriptions]
      summary: Create subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscriptionRequest'
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Subscription'
                required: [data]
        '400':
          $ref: '#/components/responses/ErrorResponse'

  /api/v1/subscriptions/update:
    post:
      tags: [Subscriptions]
      summary: Update subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSubscriptionRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Subscription'
                required: [data]
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'

  /api/v1/subscriptions/delete:
    post:
      tags: [Subscriptions]
      summary: Delete subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
              required: [id]
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '404':
          $ref: '#/components/responses/ErrorResponse'

  /api/v1/subscriptions/refresh:
    post:
      tags: [Subscriptions]
      summary: Refresh subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
              required: [id]
      responses:
        '200':
          description: Refresh result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshSubscriptionResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'

  /api/v1/nodes:
    get:
      tags: [Nodes]
      summary: List nodes
      parameters:
        - name: enabled
          in: query
          required: false
          schema:
            type: integer
            enum: [0, 1]
        - name: sub_id
          in: query
          required: false
          schema:
            type: string
        - name: q
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Node list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Node'
                required: [data]

  /api/v1/nodes/test:
    post:
      tags: [Nodes]
      summary: Test node connectivity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                node_ids:
                  type: array
                  items: { type: string }
                mode:
                  type: string
                  enum: [http, ping]
                  default: ping
              required: [node_ids]
      responses:
        '200':
          description: Test result
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        node_id: { type: string }
                        status: { type: string }
                        latency_ms: { type: integer, nullable: true }
                        error: { type: string, nullable: true }

  /api/v1/nodes/forwarding/batch:
    post:
      tags: [Nodes]
      summary: Batch set forwarding enabled for nodes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                node_ids:
                  type: array
                  items: { type: string }
                forwarding_enabled: { type: boolean }
              required: [node_ids, forwarding_enabled]
      responses:
        '200':
          description: Batch update result
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      updated: { type: integer }

  /api/v1/nodes/update:
    post:
      tags: [Nodes]
      summary: Update node
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNodeRequest'
      responses:
        '200':
          description: Updated node
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Node'
                required: [data]
        '404':
          $ref: '#/components/responses/ErrorResponse'

components:

  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'

  schemas:

    ErrorEnvelope:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorObject'
      required: [error]

    ErrorObject:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
      required: [code, message]

    Subscription:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        url: { type: string }
        type:
          type: string
          enum: [singbox]
        enabled: { type: boolean }
        auto_update_enabled: { type: boolean }
        refresh_interval_sec: { type: integer }
        used_bytes: { type: integer, nullable: true }
        total_bytes: { type: integer, nullable: true }
        remaining_bytes: { type: integer, nullable: true }
        usage_percent: { type: number, nullable: true }
        expire_at: { type: string, nullable: true }
        profile_web_page: { type: string, nullable: true }
        profile_update_interval_sec: { type: integer, nullable: true }
        created_at: { type: string }
        updated_at: { type: string }
      required: [id, name, url, type, enabled, auto_update_enabled, refresh_interval_sec, created_at, updated_at]

    CreateSubscriptionRequest:
      type: object
      properties:
        name: { type: string }
        url: { type: string }
        type:
          type: string
          enum: [singbox]
          default: singbox
        auto_update_enabled:
          type: boolean
          default: false
        refresh_interval_sec:
          type: integer
          minimum: 60
          default: 3600
      required: [url]

    UpdateSubscriptionRequest:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        url: { type: string }
        enabled: { type: boolean }
        auto_update_enabled: { type: boolean }
        refresh_interval_sec:
          type: integer
          minimum: 60
      required: [id]

    RefreshSubscriptionResponse:
      type: object
      properties:
        sub_id: { type: string }
        not_modified: { type: boolean }
        nodes_total: { type: integer }
        nodes_enabled: { type: integer }
        fetched_at: { type: string }

    Node:
      type: object
      properties:
        id: { type: string }
        sub_id: { type: string }
        tag: { type: string }
        name: { type: string }
        type: { type: string }
        enabled: { type: boolean }
        forwarding_enabled: { type: boolean }
        server: { type: string }
        server_port: { type: integer }
        network: { type: string }
        tls_enabled: { type: boolean }
        last_test_at: { type: string, nullable: true }
        last_latency_ms: { type: integer, nullable: true }
        last_test_status: { type: string, nullable: true }
        last_test_error: { type: string, nullable: true }
        created_at: { type: string }
      required: [id, sub_id, tag, name, type, enabled, forwarding_enabled, created_at]

    UpdateNodeRequest:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        enabled: { type: boolean }
        forwarding_enabled: { type: boolean }
      required: [id]

    RuntimeStatusResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            config_version: { type: integer }
            config_hash: { type: string }
            forwarding_running: { type: boolean }
            nodes_included: { type: integer }
            last_apply_duration_ms: { type: integer }
            last_apply_success_at: { type: string, nullable: true }
            last_reload_at: { type: string }
            last_reload_error: { type: string }
            ports:
              type: object
              properties:
                http: { type: integer }
                socks: { type: integer }
          required: [config_version, config_hash, forwarding_running, nodes_included, last_apply_duration_ms, ports]
      required: [data]

    RuntimeTrafficResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            sampled_at: { type: string }
            source: { type: string }
            rx_rate_bps: { type: integer }
            tx_rate_bps: { type: integer }
            rx_total_bytes: { type: integer }
            tx_total_bytes: { type: integer }
          required: [sampled_at, source, rx_rate_bps, tx_rate_bps, rx_total_bytes, tx_total_bytes]
      required: [data]

    RuntimeConnectionsResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            active_count: { type: integer }
            items:
              type: array
              items:
                type: object
                properties:
                  id: { type: string }
                  node_id: { type: string }
                  node_name: { type: string }
                  node_type: { type: string }
                  target: { type: string }
                  status: { type: string }
                  last_test_at: { type: string, nullable: true }
                  latency_ms: { type: integer, nullable: true }
                  error: { type: string, nullable: true }
                  forwarding: { type: boolean }
                  last_updated: { type: string }
                required: [id, node_id, node_name, node_type, target, status, forwarding, last_updated]
          required: [active_count, items]
      required: [data]

    RuntimeLogsResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            items:
              type: array
              items:
                type: object
                properties:
                  timestamp: { type: string }
                  level: { type: string }
                  source: { type: string }
                  message: { type: string }
                required: [timestamp, level, source, message]
          required: [items]
      required: [data]

    RuntimeProxyCheckRequest:
      type: object
      properties:
        target_url:
          type: string
          example: https://www.gstatic.com/generate_204
        timeout_ms:
          type: integer
          minimum: 500
          maximum: 30000
          default: 5000

    RuntimeProxyCheckResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            target_url: { type: string }
            checked_at: { type: string }
            http:
              $ref: '#/components/schemas/RuntimeProxyCheckItem'
            socks:
              $ref: '#/components/schemas/RuntimeProxyCheckItem'
          required: [target_url, checked_at, http, socks]
      required: [data]

    RuntimeProxyCheckItem:
      type: object
      properties:
        enabled: { type: boolean }
        proxy_url: { type: string }
        connected: { type: boolean }
        tls_ok: { type: boolean }
        status_code: { type: integer, nullable: true }
        latency_ms: { type: integer, nullable: true }
        error: { type: string, nullable: true }
        egress_ip: { type: string, nullable: true }
      required: [enabled, proxy_url, connected, tls_ok]

    ForwardingRuntimeStatusResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            running: { type: boolean }
            status:
              type: string
              enum: [running, stopped, error]
            error_message: { type: string, nullable: true }
          required: [running, status]
      required: [data]

    ForwardingSummaryResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            running: { type: boolean }
            status:
              type: string
              enum: [running, stopped, error]
            error_message: { type: string, nullable: true }
            selected_nodes_count: { type: integer }
            nodes:
              type: array
              items:
                type: object
                properties:
                  id: { type: string }
                  name: { type: string }
                  tag: { type: string }
                  type: { type: string }
                  last_status: { type: string, nullable: true }
                  last_latency_ms: { type: integer, nullable: true }
                required: [id, name, tag, type]
          required: [running, status, selected_nodes_count, nodes]
      required: [data]

    ForwardingPolicyResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            healthy_only_enabled: { type: boolean }
            max_latency_ms: { type: integer }
            allow_untested: { type: boolean }
            updated_at: { type: string, nullable: true }
          required: [healthy_only_enabled, max_latency_ms, allow_untested]
      required: [data]

    UpdateForwardingPolicyRequest:
      type: object
      properties:
        healthy_only_enabled: { type: boolean }
        max_latency_ms: { type: integer, minimum: 1, maximum: 10000 }
        allow_untested: { type: boolean }
      required: [healthy_only_enabled, max_latency_ms, allow_untested]

    RoutingSettingsResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            bypass_private_enabled: { type: boolean }
            bypass_domains:
              type: array
              items: { type: string }
            bypass_cidrs:
              type: array
              items: { type: string }
            updated_at: { type: string }
          required: [bypass_private_enabled, bypass_domains, bypass_cidrs]
      required: [data]

    RoutingSummaryResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            bypass_private_enabled: { type: boolean }
            bypass_domains_count: { type: integer }
            bypass_cidrs_count: { type: integer }
            updated_at: { type: string }
            notes:
              type: array
              items: { type: string }
          required: [bypass_private_enabled, bypass_domains_count, bypass_cidrs_count, notes]
      required: [data]

    UpdateRoutingSettingsRequest:
      type: object
      properties:
        bypass_private_enabled: { type: boolean }
        bypass_domains:
          type: array
          items: { type: string }
        bypass_cidrs:
          type: array
          items: { type: string }
      required: [bypass_private_enabled]

    RuntimePlanRequest:
      type: object
      properties:
        include_disabled_nodes:
          type: boolean
          default: false

    RuntimePlanResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            nodes_included: { type: integer }
            tags:
              type: array
              items:
                type: string
            config_hash: { type: string }
          required: [nodes_included, tags, config_hash]
      required: [data]

    RuntimeReloadRequest:
      type: object
      properties:
        force_restart:
          type: boolean
          default: true

    RuntimeReloadResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            config_version: { type: integer }
            config_hash: { type: string }
            nodes_included: { type: integer }
            restart_output: { type: string }
            reloaded_at: { type: string }
          required: [config_version, config_hash, nodes_included, reloaded_at]
      required: [data]